Move some unit tests to base unit tests directory
